Feat: 게시글 디테일 컴포넌트 쪼개기
